# openapi.yaml
openapi: 3.1.0
info:
  title: Grounded & Fruitful API
  version: 1.0.0
  description: >-
    Authorization uses canonical roles from active memberships plus organization,
    program, and resource relationships. Custom claims and route entry never grant
    tenant resource access. Domain Firestore access is server-only.
servers: [{ url: /api/v1 }]
...
paths:
...
  /participants/{id}:
    get:
      tags: [Participants]
      operationId: getParticipant
      security: [{ firebaseBearer: [] }]
      parameters:
        [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Authorized participant summary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/quarters:
    get:
      tags: [Administration]
      operationId: listAdminQuarters
      summary: List organization-scoped quarters
      description: Admins see only active membership organizations; super_admin may query any organization. Search is a case-insensitive name substring.
      security: [{ firebaseBearer: [] }]
      parameters:
        - {
            name: page,
            in: query,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 25 },
          }
        - {
            name: status,
            in: query,
            schema: { $ref: "#/components/schemas/QuarterStatus" },
          }
        - {
            name: sort,
            in: query,
            schema:
              {
                type: string,
                enum:
                  [
                    updated_desc,
                    updated_asc,
                    start_date_desc,
                    start_date_asc,
                    -updatedAt,
                    updatedAt,
                    -startDate,
                    startDate,
                  ],
                default: updated_desc,
              },
          }
        - { name: search, in: query, schema: { type: string, maxLength: 120 } }
        - {
            name: organizationId,
            in: query,
            description: Must be within the authenticated administrator's server-authorized scope.,
            schema: { type: string },
          }
      responses:
        "200":
          {
            description: Paginated quarters,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: "#/components/schemas/QuarterListResponse" },
                  },
              },
          }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
    post:
      tags: [Administration]
      operationId: createAdminQuarter
      description: Creates a draft at version 1. IDs, lifecycle, timestamps, and audit fields are server controlled. Names are unique per organization. When organizationId is omitted, the server uses the administrator's sole organization or, for a global super administrator, the deployment's sole organization. Ambiguous multi-organization requests return QUARTER_ORGANIZATION_REQUIRED.
      security: [{ firebaseBearer: [] }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/QuarterCreate" },
                  example:
                    {
                      name: Fall 2026,
                      description: Fall program quarter,
                      startDate: 2026-09-01,
                      endDate: 2026-11-30,
                      organizationId: org_123,
                    },
                },
            },
        }
      responses:
        "201":
          {
            description: Created quarter,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/QuarterResponse" } },
              },
          }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          {
            description: QUARTER_NAME_CONFLICT,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/participants:
    get:
      tags: [Administration]
      operationId: listAdminParticipants
      summary: List participants in the authenticated active workspace
      description: The active organization is server-resolved. A compatibility organizationId, when supplied, must match it.
      security: [{ firebaseBearer: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: search, in: query, schema: { type: string, minLength: 1, maxLength: 120 } }
        - { name: status, in: query, schema: { type: string, minLength: 1 } }
        - { name: teamId, in: query, schema: { type: string, minLength: 1 } }
        - { name: programId, in: query, schema: { type: string, minLength: 1 } }
        - { name: sort, in: query, schema: { type: string, enum: [updatedAt, -updatedAt], default: -updatedAt } }
        - { name: organizationId, in: query, deprecated: true, description: Compatibility value that must equal the authenticated active organization., schema: { type: string, minLength: 1 } }
      responses:
        "200": { description: Paginated participants, content: { application/json: { schema: { $ref: "#/components/schemas/ParticipantListResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
    post:
      tags: [Administration]
      operationId: createAdminParticipant
      security: [{ firebaseBearer: [] }]
      responses:
        "201": { description: Participant created }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Validation" }
  /admin/participants/{participantId}:
    parameters:
      - { name: participantId, in: path, required: true, schema: { type: string } }
    get:
      tags: [Administration]
      operationId: getAdminParticipant
      security: [{ firebaseBearer: [] }]
      responses:
        "200": { description: Participant in the authenticated active workspace }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Administration]
      operationId: updateAdminParticipant
      security: [{ firebaseBearer: [] }]
      responses:
        "200": { description: Participant updated }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
    delete:
      tags: [Administration]
      operationId: archiveAdminParticipant
      security: [{ firebaseBearer: [] }]
      responses:
        "200": { description: Participant archived }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/quarters/{quarterId}:
    parameters:
      - { $ref: "#/components/parameters/QuarterId" }
    get:
      tags: [Administration]
      operationId: getAdminQuarter
      description: Returns 404 for missing and out-of-scope records.
      security: [{ firebaseBearer: [] }]
      responses:
        "200":
          {
            description: Quarter,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/QuarterResponse" } },
              },
          }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/QuarterNotFound" }
    patch:
      tags: [Administration]
      operationId: updateAdminQuarter
      description: Updates only a draft. expectedVersion provides optimistic concurrency; a stale value returns QUARTER_VERSION_CONFLICT with HTTP 409. Lifecycle fields are never accepted here.
      security: [{ firebaseBearer: [] }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/QuarterUpdate" },
                  example:
                    {
                      name: Fall 2026,
                      description: Updated schedule,
                      expectedVersion: 1,
                    },
                },
            },
        }
      responses:
        "200":
          {
            description: Updated quarter,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/QuarterResponse" } },
              },
          }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/QuarterNotFound" }
        "409":
          {
            description: QUARTER_VERSION_CONFLICT,
            QUARTER_NAME_CONFLICT,
            or QUARTER_INVALID_TRANSITION,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "422": { $ref: "#/components/responses/Validation" }
...
components:
  parameters:
...
    QuarterId:
      {
        name: quarterId,
        in: path,
        required: true,
        schema: { type: string, minLength: 1, maxLength: 128 },
      }
...
  securitySchemes:
    {
      firebaseBearer:
        { type: http, scheme: bearer, bearerFormat: Firebase ID token },
    }
  responses:
...
    QuarterNotFound:
      {
        description: QUARTER_NOT_FOUND (also used for an out-of-scope resource),
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...
    Unauthorized:
      {
        description: Authentication required,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
    Forbidden:
      {
        description: Resource access denied,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...
    NotFound:
      {
        description: Not found,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
    Validation:
      {
        description: Validation failed,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/Error" } },
          },
      }
...
  schemas:
...
    AdminParticipant:
      type: object
      required: [id, organizationId, displayName, status]
      properties:
        id: { type: string }
        organizationId: { type: string }
        programId: { type: string }
        activeTeamId: { type: string }
        displayName: { type: string }
        status: { type: string }
        version: { type: integer, minimum: 1 }
        updatedAt: { type: string, format: date-time }
    ParticipantListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [items, pagination]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/AdminParticipant" } }
            pagination:
              type: object
              required: [page, pageSize, total, totalPages]
              properties:
                page: { type: integer, minimum: 1 }
                pageSize: { type: integer, minimum: 1, maximum: 100 }
                total: { type: integer, minimum: 0 }
                totalPages: { type: integer, minimum: 0 }
    QuarterStatus: { type: string, enum: [draft, active, closed, archived] }
    QuarterAllowedAction:
      type: string
      description: >-
        An action the authenticated caller may currently perform. The server is
        authoritative; clients must not infer permissions or lifecycle transitions.
      enum: [view, edit, activate, close, archive]
    QuarterWorkspace:
      type: object
      additionalProperties: false
      description: Authoritative minimal projection of the quarter's tenant workspace.
      required: [id, name, type]
      properties:
        id: { type: string, description: Canonical workspace identifier. }
        name:
          { type: string, description: Current server-owned workspace name. }
        type: { type: string, const: organization }
    Quarter:
      type: object
      additionalProperties: false
      required:
        [
          id,
          name,
          description,
          startDate,
          endDate,
          status,
          organizationId,
          workspace,
          allowedActions,
          createdAt,
          updatedAt,
          createdBy,
          updatedBy,
          version,
        ]
      properties:
        id: { type: string }
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, "null"], maxLength: 2000 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        status: { $ref: "#/components/schemas/QuarterStatus" }
        organizationId:
          {
            type: string,
            deprecated: true,
            description: Backward-compatible tenant identifier. New clients should use workspace.id; this field is preserved during migration.,
          }
        workspace: { $ref: "#/components/schemas/QuarterWorkspace" }
        allowedActions:
          type: array
          description: >-
            Server-calculated capabilities for this caller and lifecycle state.
            Presence in this list does not replace server enforcement on mutation.
          uniqueItems: true
          items: { $ref: "#/components/schemas/QuarterAllowedAction" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedBy: { type: string }
        version: { type: integer, minimum: 1 }
    QuarterCreate:
      type: object
      additionalProperties: false
      required: [name]
      oneOf:
        - required: [startDate, endDate]
        - required: [startsOn, endsOn]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, "null"], maxLength: 2000 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        startsOn:
          {
            type: string,
            format: date,
            deprecated: true,
            description: Compatibility alias for startDate. Do not send both naming styles.,
          }
        endsOn:
          {
            type: string,
            format: date,
            deprecated: true,
            description: Compatibility alias for endDate. Do not send both naming styles.,
          }
        organizationId:
          {
            type: string,
            description: Optional when the target is unambiguous (the administrator's sole organization or the sole deployment organization for a global super administrator); otherwise required.,
          }
    QuarterUpdate:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      minProperties: 2
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, "null"], maxLength: 2000 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        expectedVersion: { type: integer, minimum: 1 }
...
    QuarterResponse:
      type: object
      required: [data]
      properties: { data: { $ref: "#/components/schemas/Quarter" } }
    QuarterListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              { type: array, items: { $ref: "#/components/schemas/Quarter" } }
            pagination:
              type: object
              required: [page, pageSize, total, totalPages]
              properties:
                {
                  page: { type: integer },
                  pageSize: { type: integer },
                  total: { type: integer },
                  totalPages: { type: integer },
                }
...
    Error:
      {
        type: object,
        required: [error],
        properties:
          {
            error:
              {
                type: object,
                required: [code, message, requestId],
                properties:
                  {
                    code: { type: string },
                    message: { type: string },
                    requestId: { type: string },
                    fieldErrors: { type: object, additionalProperties: true },
                  },
              },
          },
      }
...